feat: fish process to charge order sync to the system and update tables

+++ common-module/src/main/java/com/cumpleanos/core/models/entities/Dfactura.java
@@ -51,7 +51,7 @@ public class Dfactura {
 
     @NotNull
     @Column(name = "DFAC_CANTIDAD", nullable = false, precision = 17, scale = 4)
-    private Long cantidad;
+    private Integer cantidad;
 
     @NotNull
     @Column(name = "DFAC_CANAPR", nullable = false, precision = 17, scale = 4)
